requirements.txt
# fl-privacy-project requirements
# ================================
# Install with: pip install -r requirements.txt
#
# Tested on Python 3.9 – 3.11
# PyTorch: CPU build by default.
# For CUDA, install PyTorch separately:
#   pip install torch --index-url https://download.pytorch.org/whl/cu118

# ── Core deep learning ────────────────────────────────────────────
torch>=2.0.0,<3.0.0
torchvision>=0.15.0          # only needed if extending to image datasets

# ── ML utilities ─────────────────────────────────────────────────
scikit-learn>=1.3.0
numpy>=1.24.0

# ── Config ───────────────────────────────────────────────────────
PyYAML>=6.0

# ── Data / analysis ──────────────────────────────────────────────
pandas>=2.0.0                # ExperimentRegistry.to_dataframe()

# ── Visualisation ────────────────────────────────────────────────
matplotlib>=3.7.0

# ── Testing ──────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0            # coverage reports

# ── Development (optional) ───────────────────────────────────────
# jupyter>=1.0.0             # for notebooks/analysis.ipynb
# ipykernel>=6.0.0
# black>=23.0.0              # code formatting
# ruff>=0.1.0                # linting
